refactor(storage): move the Disk layout inside the storage boundary
`workspace/disk/paths.ts` described two unrelated things: where the Disk
backend keeps a Space, and where the workspace keeps things that have
nothing to do with a backend. Sorting them by the §12.5.2 test — is this
still useful once the structured backend is SQLite? — splits cleanly.

The Disk record and blob layout moves to `backends/disk/layout.ts`,
together with `canvas-dirs.ts` and `name-index.ts`. Those two had zero
consumers outside `storage/` already; they were filed under a `disk`
segment in the workspace module while being, in fact, Disk
structured-backend state built by parsing `space.json`.

What stays behind is the residue that survives the test: `setting/`, user
skills, user memory, and the per-Space state other domains own. Those
anchor on a new `spaceDirectory()` capability exported by the composition
root, so they no longer resolve through the Disk name index — the Space
counterpart to `BlobScope.materialize()` (§12.5.4).

The shim importer list shrinks by fourteen: most consumers turned out to
want workspace-owned paths and now say so directly. Five entries join it
as relocations of couplings that already existed, and the four remaining
storage-owned reads outside the boundary — `chatDir`, `nodesDir`,
`changesPath`, `artifactsDir` — are recorded in the boundary test rather
than hidden, for step 5 to remove.

Behavior is unchanged: every path resolves to the byte-identical location,
and no Disk test expectation moved.

Phase 4.5, step 2 of 6 (proposal §12.5.5).
