+++ apps/server/src/modules/agent/conversation/prompt/debug-prompt.ts
@@ -17,7 +17,8 @@
 import { appendFileSync } from 'node:fs';
 
 import { mkdirp } from '../../../../utils/fs.js';
-import { chatDir, chatPromptLogPath } from '../../../storage/paths.js';
+import { chatDir } from '../../../storage/paths.js';
+import { chatPromptLogPath } from '../../../workspace/disk/paths.js';
 
 import type { Context } from '@earendil-works/pi-ai';
 import type { FastifyBaseLogger } from 'fastify';
