+++ apps/server/src/modules/agent/memory/trigger.ts
@@ -26,11 +26,11 @@ import { existsSync } from 'node:fs';
 
 import { atomicWriteJson, mkdirp, readJson } from '../../../utils/fs.js';
 import { createKeyedMutex } from '../../../utils/keyed-mutex.js';
+import { spaceDirectory } from '../../storage/index.js';
 import {
   memoryStatePath,
   canvasMemoryDir,
-  canvasRoot,
-} from '../../storage/paths.js';
+} from '../../workspace/disk/paths.js';
 
 /** Op-count threshold that triggers a memory analysis pass. */
 export const OP_THRESHOLD = 50;
@@ -86,7 +86,7 @@ export function writeMemoryState(canvasId: string, state: MemoryState): void {
   // file. Same hazard for any in-flight memory worker that calls
   // `markAnalyzed` post-delete. Skip the write when the canvas root
   // is gone; losing one bookkeeping write is harmless.
-  if (!existsSync(canvasRoot(canvasId))) return;
+  if (!existsSync(spaceDirectory(canvasId))) return;
   mkdirp(canvasMemoryDir(canvasId));
   atomicWriteJson(memoryStatePath(canvasId), state);
 }
