+++ apps/server/src/modules/agent/tools/handlers/fs-write.ts
@@ -36,7 +36,7 @@ import {
   canvasMemoryDir,
   settingDir,
   userSkillsDir,
-} from '../../../storage/paths.js';
+} from '../../../workspace/disk/paths.js';
 import {
   resolveLongTermPath,
   resolveUserSkillPath,
