+++ apps/server/src/modules/storage/backends/disk/cache-boundaries.test.ts
@@ -14,6 +14,8 @@ import path from 'node:path';
 
 import { afterEach, describe, expect, it, vi } from 'vitest';
 
+import { refreshCanvasDirIndex, registerCanvasDir } from './canvas-dirs.js';
+import { canvasRoot, SPACE_JSON_FILENAME } from './layout.js';
 import {
   forgetCanvasStore,
   getCanvasStore,
@@ -22,14 +24,6 @@ import {
 import { NODE_TOMBSTONE_TTL_MS } from './legacy/node-tombstones.js';
 import { DiskStructuredStore } from './structured-store.js';
 import { toSafeFilename } from '../../../../utils/naming.js';
-import {
-  refreshCanvasDirIndex,
-  registerCanvasDir,
-} from '../../../workspace/disk/canvas-dirs.js';
-import {
-  canvasRoot,
-  SPACE_JSON_FILENAME,
-} from '../../../workspace/disk/paths.js';
 import { setWorkspacePath } from '../../../workspace.js';
 
 import type {
