+++ apps/server/src/modules/storage/backends/disk/structured-store.test.ts
@@ -13,14 +13,14 @@ vi.mock('../../../workspace.js', () => ({
   getWorkspacePath: () => workspaceState.path,
 }));
 
+import { refreshCanvasDirIndex } from './canvas-dirs.js';
+import { tasksPath } from './layout.js';
 import {
   getCanvasStore,
   resetStorageCache,
 } from './legacy/canvas-store-cache.js';
 import { DiskStructuredStore } from './structured-store.js';
 import { toSafeFilename } from '../../../../utils/naming.js';
-import { refreshCanvasDirIndex } from '../../../workspace/disk/canvas-dirs.js';
-import { tasksPath } from '../../../workspace/disk/paths.js';
 import { describeSpaceLogsContract } from '../../ports/contracts/space-logs.contract.js';
 import { describeStructuredStoreContract } from '../../ports/contracts/structured-store.contract.js';
