+++ apps/server/src/modules/storage/compatibility/canvas.ts
@@ -22,17 +22,17 @@ import path from 'node:path';
 
 import { atomicWriteJson, mkdirp, sanitizeId } from '../../../utils/fs.js';
 import { toSafeFilename } from '../../../utils/naming.js';
+import { getWorkspacePath } from '../../workspace.js';
 import {
   listCanvasDirEntries,
   refreshCanvasDirIndex,
   registerCanvasDir,
   suggestCanvasDir,
-} from '../../workspace/disk/canvas-dirs.js';
+} from '../backends/disk/canvas-dirs.js';
 import {
   canvasJsonPath,
   SPACE_JSON_FILENAME,
-} from '../../workspace/disk/paths.js';
-import { getWorkspacePath } from '../../workspace.js';
+} from '../backends/disk/layout.js';
 import { getCanvasStore } from '../backends/disk/legacy/canvas-store-cache.js';
 import { deleteSpace } from '../storage.js';
